+++ onnx/helper.py
@@ -351,7 +351,11 @@ def float32_to_bfloat16(fval: float, truncate: bool = False) -> int:
 
 
 def float32_to_float8e4m3(  # pylint: disable=too-many-statements
-    fval: float, scale: float = 1.0, fn: bool = True, uz: bool = False
+    fval: float,
+    scale: float = 1.0,
+    fn: bool = True,
+    uz: bool = False,
+    saturate: bool = True,
 ) -> int:
     """
     Convert a float32 value to a float8, e4m3 (as int).
@@ -360,6 +364,9 @@ def float32_to_float8e4m3(  # pylint: disable=too-many-statements
     :param scale: scale, divide *fval* by *scale* before casting it
     :param fn: no infinite values
     :param uz: no negative zero
+    :param saturate: if True, any value out of range included inf becomes the maximum value,
+        otherwise, it becomes NaN. The description of operator Cast fully describes the
+        differences.
     :return: converted float
 
     See :ref:`onnx-detail-float8` for technical details.
@@ -375,6 +382,8 @@ def float32_to_float8e4m3(  # pylint: disable=too-many-statements
         if (b & 0x7FC00000) == 0x7FC00000:
             return 0x80
         if np.isinf(x):
+            if saturate:
+                return ret | 126
             return 0x80
         e = (b & 0x7F800000) >> 23  # exponent
         m = b & 0x007FFFFF  # mantissa
@@ -402,16 +411,26 @@ def float32_to_float8e4m3(  # pylint: disable=too-many-statements
                 else:
                     ret |= ex << 3
                     ret |= m >> 20
-                if (m & 0x80000) and (ret & 0x7F) < 0x7F:
-                    # rounding
-                    ret += 1
-            else:
+                if m & 0x80000:
+                    if (ret & 0x7F) < 0x7F:
+                        # rounding
+                        ret += 1
+                    elif not saturate:
+                        return 0x80
+            elif saturate:
                 ret |= 0x7F  # 01111110
+            else:
+                ret = 0x80
+        elif m == 0:
+            # -0
+            ret = 0
         return int(ret)
     else:
         if (b & 0x7FC00000) == 0x7FC00000:
             return 0x7F | ret
         if np.isinf(x):
+            if saturate:
+                return ret | 126
             return 0x7F | ret
         e = (b & 0x7F800000) >> 23  # exponent
         m = b & 0x007FFFFF  # mantissa
@@ -441,16 +460,25 @@ def float32_to_float8e4m3(  # pylint: disable=too-many-statements
                     ret |= m >> 20
                     if (ret & 0x7F) == 0x7F:
                         ret &= 0xFE
-                if (m & 0x80000) and (ret & 0x7F) < 0x7E:
-                    # rounding
-                    ret += 1
-            else:
+                if m & 0x80000:
+                    if (ret & 0x7F) < 0x7E:
+                        # rounding
+                        ret += 1
+                    elif not saturate:
+                        ret |= 0x7F
+            elif saturate:
                 ret |= 126  # 01111110
+            else:
+                ret |= 0x7F
         return int(ret)
 
 
 def float32_to_float8e5m2(  # pylint: disable=too-many-statements
-    fval: float, scale: float = 1.0, fn: bool = False, uz: bool = False
+    fval: float,
+    scale: float = 1.0,
+    fn: bool = False,
+    uz: bool = False,
+    saturate: bool = True,
 ) -> int:
     """
     Convert a float32 value to a float8, e5m2 (as int).
@@ -459,6 +487,9 @@ def float32_to_float8e5m2(  # pylint: disable=too-many-statements
     :param scale: scale, divide *fval* by *scale* before casting it
     :param fn: no infinite values
     :param uz: no negative zero
+    :param saturate: if True, any value out of range included inf becomes the maximum value,
+        otherwise, it becomes NaN. The description of operator Cast fully describes the
+        differences.
     :return: converted float
     """
     x = fval / scale
@@ -467,6 +498,11 @@ def float32_to_float8e5m2(  # pylint: disable=too-many-statements
 
     if fn and uz:
         if (b & 0x7FC00000) == 0x7FC00000:
+            return 0x80
+        if (b & 0x7FFFFFFF) == 0x7F800000:
+            # inf
+            if saturate:
+                return ret | 0x7F
             return 0x80
         e = (b & 0x7F800000) >> 23  # exponent
         m = b & 0x007FFFFF  # mantissa
@@ -491,17 +527,29 @@ def float32_to_float8e5m2(  # pylint: disable=too-many-statements
                 ex = e - 111  # 127 - 16
                 ret |= ex << 2
                 ret |= m >> 21
-                if (m & 0x100000) and (ret & 0x7F) < 0x7F:
-                    # rounding
-                    ret += 1
+                if m & 0x100000:
+                    if (ret & 0x7F) < 0x7F:
+                        # rounding
+                        ret += 1
+                    elif not saturate:
+                        ret = 0x80
             elif e == 255 and m == 0:  # inf
-                return 0x80
-            else:
+                ret = 0x80
+            elif saturate:
                 ret |= 0x7F  # last possible number
+            else:
+                ret = 0x80
+        elif m == 0:
+            # -0
+            ret = 0
         return int(ret)
     elif not fn and not uz:
         if (b & 0x7FC00000) == 0x7FC00000:
             return 0x7F | ret
+        if np.isinf(x):
+            if saturate:
+                return 0x7B | ret
+            return 0x7C | ret
         e = (b & 0x7F800000) >> 23  # exponent
         m = b & 0x007FFFFF  # mantissa
 
@@ -525,13 +573,18 @@ def float32_to_float8e5m2(  # pylint: disable=too-many-statements
                 ex = e - 112  # 127 - 15
                 ret |= ex << 2
                 ret |= m >> 21
-                if (m & 0x100000) and (ret & 0x7F) < 0x7B:
-                    # rounding
-                    ret += 1
-            elif e == 255 and m == 0:  # inf
-                ret |= 124
+                if m & 0x100000:
+                    if (ret & 0x7F) < 0x7B:
+                        # rounding
+                        ret += 1
+                    elif saturate:
+                        ret |= 0x7B
+                    else:
+                        ret |= 0x7C
+            elif saturate:
+                ret |= 0x7B
             else:
-                ret |= 123
+                ret |= 0x7C
         return int(ret)
     else:
         raise NotImplementedError("fn and uz must be both False or True.")
@@ -754,48 +807,30 @@ def make_optional(
     return optional
 
 
-def _to_bytes_or_false(val: Union[str, bytes]) -> Union[bytes, bool]:
-    """An internal graph to convert the input to a bytes or to False.
-
-    The criteria for conversion is as follows and should be python 2 and 3
-    compatible:
-    - If val is py2 str or py3 bytes: return bytes
-    - If val is py2 unicode or py3 str: return val.decode('utf-8')
-    - Otherwise, return False
-    """
-    if isinstance(val, bytes):
-        return val
-    try:
-        return val.encode("utf-8")
-    except AttributeError:
-        return False
+def _to_bytes(value: Union[str, bytes]) -> bytes:
+    """Coerce a string (or bytes) value into UTF-8 bytes."""
+    return value if isinstance(value, bytes) else value.encode("utf-8")
 
 
 def make_attribute(  # pylint: disable=too-many-statements
     key: str, value: Any, doc_string: Optional[str] = None
-) -> AttributeProto:  # pylint: disable=too-many-statements
+) -> AttributeProto:
     """Makes an AttributeProto based on the value type."""
     attr = AttributeProto()
     attr.name = key
     if doc_string:
         attr.doc_string = doc_string
 
-    is_iterable = isinstance(value, collections.abc.Iterable)
-    bytes_or_false = _to_bytes_or_false(value)
-    # First, singular cases
-    # float
-    if isinstance(value, float):
-        attr.f = value
-        attr.type = AttributeProto.FLOAT
-    # integer
-    elif isinstance(value, numbers.Integral):
-        attr.i = cast(int, value)
+    # Singular cases
+    if isinstance(value, numbers.Integral):
+        attr.i = int(value)
         attr.type = AttributeProto.INT
-    # string
-    elif bytes_or_false is not False:
-        if not isinstance(bytes_or_false, bytes):
-            raise TypeError(f"bytes_or_false must be an instance of {bytes}.")
-        attr.s = bytes_or_false
+    elif isinstance(value, numbers.Real):
+        attr.f = float(value)
+        attr.type = AttributeProto.FLOAT
+    elif isinstance(value, (str, bytes)):
+        # Encode strings into utf-8
+        attr.s = _to_bytes(value)
         attr.type = AttributeProto.STRING
     elif isinstance(value, TensorProto):
         attr.t.CopyFrom(value)
@@ -809,40 +844,37 @@ def make_attribute(  # pylint: disable=too-many-statements
     elif isinstance(value, TypeProto):
         attr.tp.CopyFrom(value)
         attr.type = AttributeProto.TYPE_PROTO
-    # third, iterable cases
-    elif is_iterable:
-        byte_array = [_to_bytes_or_false(v) for v in value]
-        if all(isinstance(v, numbers.Integral) for v in value):
-            # Turn np.int32/64 into Python built-in int.
-            attr.ints.extend(int(v) for v in value)
+    # Iterable cases
+    elif isinstance(value, collections.abc.Iterable):
+        value = list(value)
+        types = {type(v) for v in value}
+        if all(issubclass(t, numbers.Integral) for t in types):
+            attr.ints.extend(value)
             attr.type = AttributeProto.INTS
-        elif all(isinstance(v, numbers.Real) for v in value):
-            # Since ints and floats are members of Real, this allows a mix of ints and floats
-            # (and converts the ints to floats).
-            attr.floats.extend(float(v) for v in value)
+        elif all(issubclass(t, numbers.Real) for t in types):
+            attr.floats.extend(value)
             attr.type = AttributeProto.FLOATS
-        elif all(map(lambda bytes_or_false: bytes_or_false is not False, byte_array)):
-            attr.strings.extend(cast(List[bytes], byte_array))
+        elif all(issubclass(t, (str, bytes)) for t in types):
+            attr.strings.extend(_to_bytes(v) for v in value)
             attr.type = AttributeProto.STRINGS
-        elif all(isinstance(v, TensorProto) for v in value):
+        elif all(issubclass(t, TensorProto) for t in types):
             attr.tensors.extend(value)
             attr.type = AttributeProto.TENSORS
-        elif all(isinstance(v, SparseTensorProto) for v in value):
+        elif all(issubclass(t, SparseTensorProto) for t in types):
             attr.sparse_tensors.extend(value)
             attr.type = AttributeProto.SPARSE_TENSORS
-        elif all(isinstance(v, GraphProto) for v in value):
+        elif all(issubclass(t, GraphProto) for t in types):
             attr.graphs.extend(value)
             attr.type = AttributeProto.GRAPHS
-        elif all(isinstance(tp, TypeProto) for tp in value):
+        elif all(issubclass(t, TypeProto) for t in types):
             attr.type_protos.extend(value)
             attr.type = AttributeProto.TYPE_PROTOS
         else:
             raise ValueError(
-                "You passed in an iterable attribute but I cannot figure out "
-                "its applicable type."
+                "Could not infer the attribute type from the elements of the passed Iterable value."
             )
     else:
-        raise TypeError(f'value "{value}" is not valid attribute data type.')
+        raise TypeError(f"'{value}' is not an accepted attribute value.")
     return attr
 
 
@@ -890,6 +922,15 @@ def get_attribute_value(attr: AttributeProto) -> Any:
     if attr.type == AttributeProto.TYPE_PROTOS:
         return list(attr.type_protos)
     raise ValueError(f"Unsupported ONNX attribute: {attr}")
+
+
+def get_node_attr_value(node: NodeProto, attr_name: str) -> Any:
+    matching = [x for x in node.attribute if x.name == attr_name]
+    if len(matching) > 1:
+        raise ValueError(f"Node has multiple attributes with name {attr_name}")
+    if len(matching) < 1:
+        raise ValueError(f"Node has no attribute with name {attr_name}")
+    return get_attribute_value(matching[0])
 
 
 def make_empty_tensor_value_info(name: str) -> ValueInfoProto: