+++ onnx/reference/ops/_op_list.py
@@ -1,3 +1,5 @@
+# Copyright (c) ONNX Project Contributors
+
 # SPDX-License-Identifier: Apache-2.0
 # pylint: disable=C0415,R0912,R0913,R0914,R0915,W0611,W0603
 """
@@ -55,8 +57,8 @@ from onnx.reference.ops.op_bitwise_not import BitwiseNot
 from onnx.reference.ops.op_bitwise_or import BitwiseOr
 from onnx.reference.ops.op_bitwise_xor import BitwiseXor
 from onnx.reference.ops.op_blackman_window import BlackmanWindow
-from onnx.reference.ops.op_cast import Cast
-from onnx.reference.ops.op_cast_like import CastLike
+from onnx.reference.ops.op_cast import Cast_1, Cast_19
+from onnx.reference.ops.op_cast_like import CastLike_15, CastLike_19
 from onnx.reference.ops.op_ceil import Ceil
 from onnx.reference.ops.op_celu import Celu
 from onnx.reference.ops.op_center_crop_pad import CenterCropPad
@@ -78,6 +80,7 @@ from onnx.reference.ops.op_conv_transpose import ConvTranspose
 from onnx.reference.ops.op_cos import Cos
 from onnx.reference.ops.op_cosh import Cosh
 from onnx.reference.ops.op_cum_sum import CumSum
+from onnx.reference.ops.op_deform_conv import DeformConv
 from onnx.reference.ops.op_depth_to_space import DepthToSpace
 from onnx.reference.ops.op_dequantize_linear import DequantizeLinear
 from onnx.reference.ops.op_det import Det
@@ -149,7 +152,7 @@ from onnx.reference.ops.op_pow import Pow
 from onnx.reference.ops.op_prelu import PRelu
 from onnx.reference.ops.op_qlinear_conv import QLinearConv
 from onnx.reference.ops.op_qlinear_matmul import QLinearMatMul
-from onnx.reference.ops.op_quantize_linear import QuantizeLinear
+from onnx.reference.ops.op_quantize_linear import QuantizeLinear_10, QuantizeLinear_19
 from onnx.reference.ops.op_random_normal import RandomNormal
 from onnx.reference.ops.op_random_normal_like import RandomNormalLike
 from onnx.reference.ops.op_random_uniform import RandomUniform